requirements.txt
fastapi
uvicorn
groq
pydantic
python-multipart
pdfplumber
httpx
python-dotenv